Move symbol and .symbol.mk to the scaffold folder
These are the templated ones that referenced TARGET.cm and TARGET.mlb.
We're going to use local symbol to build the global symbol command in
SML, so we'll soon have a duplciate of these scripts, modulo the target
names.

The ones with `target=TARGET` now live in scaffold after this commit.