Fix bug with arrays on bash 3
https://stackoverflow.com/questions/7577052/

Basically, an empty array in bash is considered to be unset, so if you
do "${foo[@]}" in conjunction with `set -u`, you'll get an unbound
variable error and the script will crash.

I never caught this because I was only testing against bash 4 locally.
See the Stack Overflow script for the workaround.